Refactor: relocate schema-aware passes (JsonPathAnnotator, RelationIdRewriter) into lib/passes (#14)
* Initial plan

* Refactor: move schema-aware passes (JsonPathAnnotator, RelationIdRewriter) into lib/passes

Agent-Logs-Url: https://github.com/TekuConcept/casl-bridge/sessions/f16310b3-c4a3-4d89-85fe-a3921601c644


* Moved tests to correct folder + removed file stubs

---------
