+++ lib/condition/index.ts
@@ -1,8 +1,6 @@
 export * from './base-condition'
-export * from './json-path-annotator'
 export * from './literal-condition'
 export * from './mongo-query'
 export * from './primitive-condition'
-export * from './relation-id-rewriter'
 export * from './scoped-condition'
 export * from './types'
