+++ lib/passes/json-path-annotator.spec.ts
