+++ lib/passes/relation-id-rewriter.spec.ts
@@ -1,14 +1,14 @@
 import 'mocha'
 import { expect } from 'chai'
+import { ScopedCondition } from '../condition/scoped-condition'
+import { PrimitiveCondition } from '../condition/primitive-condition'
+import { LiteralCondition } from '../condition/literal-condition'
+import { PrimOp, ScopeOp } from '../condition/types'
 import {
     RelationIdRewriter,
     RelationIdMeta,
     RelationMetaProvider,
 } from './relation-id-rewriter'
-import { ScopedCondition } from './scoped-condition'
-import { PrimitiveCondition } from './primitive-condition'
-import { LiteralCondition } from './literal-condition'
-import { PrimOp, ScopeOp } from './types'
 
 // ─────────────────────────────────────────────────────────────────────────────
 // Test helpers
@@ -382,7 +382,7 @@ describe('RelationIdRewriter', () => {
                 const p = prim('id', 1)
                 const rw = new RelationIdRewriter(provider({}))
                 const result = rw.apply(p)
-                expect(result).to.equal(p)
+                expect(result.tree).to.equal(p)
             })
 
             it('should handle multiple join scopes at the same level', () => {
